+++ cmd/github-actions-adapter/serve/comments.go
@@ -11,14 +11,14 @@ import (
 
 // commentResultOnPatch adds a patch-revision comment with the results of the GitHub workflows.
 func (gas *GitHubActionsServer) commentOnPatch(ctx context.Context,
-	brokerRequestMessage *broker.RequestMessage, commentMessage string) error {
+	brokerRequestMessage *broker.RequestMessage, commentMessage string, append bool) error {
 	if len(brokerRequestMessage.PatchEvent.Patch.Revisions) == 0 {
 		gas.App.Logger.Warn("could not comment on patch", "error", "no revision found in patch")
 		return errors.New("no revision found in patch")
 	}
 	revision := brokerRequestMessage.PatchEvent.Patch.Revisions[len(brokerRequestMessage.PatchEvent.Patch.Revisions)-1]
 	err := gas.Radicle.Comment(ctx, brokerRequestMessage.Repo, brokerRequestMessage.PatchEvent.Patch.ID, revision.ID,
-		commentMessage)
+		commentMessage, append)
 	if err != nil {
 		gas.App.Logger.Warn("could not comment on patch", "content", commentMessage, "patch_id",
 			brokerRequestMessage.PatchEvent.Patch.ID, "revision_id", revision.ID, "error", err.Error())